fix(services): move data files from feature module to `app/data`
