+++ live-editing/configs/TransactionsConfigGenerator.ts
@@ -3,7 +3,7 @@ IgxCardModule,
 IgxIconModule,
 IgxListModule,
 IgxTransactionService} from 'igniteui-angular';
-import {AppModuleConfig, Config, IConfigGenerator} from 'igniteui-live-editing'
+import {AppModuleConfig, Config, IConfigGenerator} from 'igniteui-live-editing';
 export class TransactionsConfigGenerator implements IConfigGenerator {
     public additionalImports = {
         TransactionBasePipe: '../../src/app/services/transaction/pipes/transaction-base.pipe'
@@ -13,8 +13,8 @@ export class TransactionsConfigGenerator implements IConfigGenerator {
 
         configs.push(new Config({
             component: 'TransactionBaseComponent',
-            additionalFiles: ["/src/app/services/transaction/data.ts",
-                "/src/app/services/transaction/pipes/transaction-base.pipe.ts"],
+            additionalFiles: ['/src/app/data/wishlist-data.ts',
+                '/src/app/services/transaction/pipes/transaction-base.pipe.ts'],
             appModuleConfig: new AppModuleConfig({
                 imports: ['IgxListModule', 'IgxCardModule', 'IgxIconModule', 'IgxButtonModule',
                     'IgxTransactionService', 'TransactionBaseComponent', 'TransactionBasePipe'],
@@ -22,7 +22,7 @@ export class TransactionsConfigGenerator implements IConfigGenerator {
                 ngImports: ['IgxListModule', 'IgxCardModule', 'IgxButtonModule', 'IgxIconModule'],
                 ngProviders: ['IgxTransactionService']
             }),
-            shortenComponentPathBy: "services/transaction-base"
+            shortenComponentPathBy: 'services/transaction-base'
         }));
 
         return configs;
