+++ src/app/services/localization-samples/localization-sample-1/localization-sample-1.component.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { IgxGridComponent } from 'igniteui-angular';
-import { DATA } from '../../data/nwindData';
+import { DATA } from '../../../data/nwindData';
 
 @Component({
     selector: 'app-localization-sample-1',
