+++ src/app/services/localization-samples/localization-sample-3/localization-sample-3.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { IgxGridComponent } from 'igniteui-angular';
-import { DATA } from '../../data/nwindData';
+import { DATA } from '../../../data/nwindData';
 
 @Component({
     selector: 'app-localization-sample-3',
